fix Delta's git patch navigation
Need navigate = true so that Delta will add the •Δ symbols to search on.
Need to wrap those symbols in parenthesis lest the ^ only apply to the first symbol.
I had this commit ready on the delta testing branch but forgot to push it and merged https://github.com/PatrickF1/dotfiles/pull/25 before I realized.